TST: Reorganizing corrections tests.
